rename periphery to peripheral
