+++ Src/modules/cyphal/feedback/feedback.cpp
@@ -8,7 +8,7 @@
 #include <algorithm>
 #include "cyphal.hpp"
 #include "params.hpp"
-#include "periphery/pwm/pwm.hpp"
+#include "peripheral/pwm/pwm.hpp"
 
 REGISTER_MODULE(FeedbackModule)
