+++ Src/peripheral/adc/adc.hpp
@@ -5,7 +5,7 @@
 #define SRC_PERIPHERY_ADC_ADC_HPP_
 
 #include <stdint.h>
-#include "periphery/temperature_sensor/temperature_sensor.hpp"
+#include "peripheral/temperature_sensor/temperature_sensor.hpp"
 #ifdef __cplusplus
 extern "C" {
 #endif
