+++ Src/platform/stm32/i2c/i2c.cpp
@@ -4,7 +4,7 @@
  * Author: Dmitry Ponomarev <ponomarevda96@gmail.com>
  */
 
-#include "periphery/i2c/i2c.hpp"
+#include "peripheral/i2c/i2c.hpp"
 #include <array>
 #include "i2c.h"
