+++ Src/platform/ubuntu/iwdg.cpp
@@ -4,7 +4,7 @@
  * Author: Dmitry Ponomarev <ponomarevda96@gmail.com>
  */
 
-#include "periphery/iwdg/iwdg.hpp"
+#include "peripheral/iwdg/iwdg.hpp"
 
 void WatchdogPeriphery::refresh() {
     // do nothing
